requirements.txt
torch_geometric
pyg_lib
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv
https://data.pyg.org/whl/torch-2.2.1%2Bcpu.html
